Move transform kit to BasePluginTest
